Add files via upload
The basic structure of how we can store the black holes in an array. There is no OOP in Mathematica, so effectively each entry in the blackHoleArray is a black hole, each entry being a list of 5 qualities of that black hole, the xPos, yPos, xVel, yVel, and mass of that black hole. This program also gets that input from the user, so I think it'd be a good spring board for whatever we do next. I started making the outline of determining the change in velocity and position for each iteration (it's the commented out stuff).